- add PastAppointmentCard widget
